# modules/dxp/apps/commerce/headless/headless-commerce-shop-by-diagram-impl/rest-openapi.yaml
components:
...
    schemas:
...
        DiagramEntry:
            properties:
                diagram:
                    example: false
                    type: boolean
                expando:
                    additionalProperties: true
                    type: object
                id:
                    example: 33130
                    format: int64
                    minimum: 0
                    type: integer
                productId:
                    example: 33131
                    format: int64
                    minimum: 0
                    type: integer
                quantity:
                    example: 1
                    format: int32
                    minimum: 0
                    type: integer
                sequence:
                    example: 1
                    type: string
                sku:
                    example: SKU01
                    type: string
                skuId:
                    example: 33131
                    format: int64
                    minimum: 0
                    type: integer
            type: object
        Error:
            properties:
                errorCode:
                    description:
                        Internal error code mapping
                    example: 996
                    readOnly: true
                    type: integer
                errorDescription:
                    example: Unable to find currency. Currency code should be expressed with
                        3-letter ISO 4217 format.
                    readOnly: true
                    type: string
                message:
                    example: No CommerceCurrency exists with the key {catalogId=41811, code=US
                        Dollar}
                    readOnly: true
                    type: string
                status:
                    description:
                        HTTP Status code
                    example: 404
                    readOnly: true
                    type: integer
            required:
                - errorCode
                - errorDescription
                - message
                - status
            type: object
...
    securitySchemes:
        BasicAuth:
            scheme: basic
            type: http
        BearerAuth:
            scheme: bearer
            type: http
        OAuth2:
            flows:
                authorizationCode:
                    authorizationUrl: /oauth/authorize
                    scopes:
                        HeadlessCommerceAdminCatalog.admin: Grants access to admin operations
                        HeadlessCommerceAdminCatalog.read: Grants read access
                        HeadlessCommerceAdminCatalog.write: Grants write access
                    tokenUrl: /oauth/token
            type: oauth2
info:
    contact:
        email: team-commerce@liferay.com
        name: Commerce Team
    description:
        "Liferay Commerce Admin Catalog API. A Java client JAR is available for use with the group ID 'com.liferay',
        artifact ID 'com.liferay.headless.commerce.admin.catalog.client', and version '4.0.3'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: Liferay Commerce Admin Catalog API
    version: v1.0
openapi: 3.0.0
paths:
    /diagramEntries/{diagramEntryId}:
        delete:
            operationId: deleteDiagramEntry
            parameters:
                - in: path
                  explode: false
                  name: diagramEntryId
                  required: true
                  schema:
                      description:
                          The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Deletes a Diagram entry by its ID.
            tags:
                - DiagramEntry
        patch:
            operationId: patchDiagramEntry
            parameters:
                - in: path
                  explode: false
                  name: diagramEntryId
                  required: true
                  schema:
                      description:
                          The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DiagramEntry"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DiagramEntry"
                required: true
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DiagramEntry"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DiagramEntry"
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Updates a Diagram entry by its ID.
            tags:
                - DiagramEntry
...
    /products/by-externalReferenceCode/{externalReferenceCode}/diagramEntries:
        get:
            operationId: getProductByExternalReferenceCodeDiagramEntriesPage
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
                - in: query
                  explode: true
                  name: page
                  required: false
                  schema:
                      type: integer
                  style: form
                - in: query
                  explode: true
                  name: pageSize
                  required: false
                  schema:
                      type: integer
                  style: form
                - in: query
                  name: search
                  required: false
                  schema:
                      type: string
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DiagramEntry"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DiagramEntry"
                                type: array
                    description:
                        Successful operation
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.read
            summary: Gets a list of Product by external reference code.
            tags:
                - DiagramEntry
        post:
            operationId: postProductByExternalReferenceCodeDiagramEntry
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DiagramEntry"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DiagramEntry"
                required: true
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DiagramEntry"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DiagramEntry"
                    description:
                        Created
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Creates a diagram entry related to a product by external reference code.
            tags:
                - DiagramEntry
...
servers:
    - description:
          SwaggerHub API Auto Mocking
      url: https://virtserver.swaggerhub.com/liferayinc/commerce-admin-catalog/v1.0